\name{BCMETA2-package}
\alias{BCMETA2-package}
\alias{BCMETA2}
\docType{package}
\title{
Meta-analysis of gene expression and clinical data in breast cancer
}
\description{
Computational pipeline that generate several useful analysis in breast cancer studies. The user may choose from a configuration file which dataset he wants to include as a training set. Base on those datasets the pipeline will then predict how well a specify geneset will perform in prognostic analysis.
}
\details{
\tabular{ll}{
Package: \tab BCMETA2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-25\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Pierre-Olivier Bachant-Winner
Benjamin Haibe-Kains

Maintainer: benjamin.haibe.kains@ircm.qc.ca
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ Survival analysis }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
